requirements.txt
blinker==1.9.0
click==8.1.8
colorama==0.4.6
Flask==3.1.0
gunicorn==21.2.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
pillow==11.1.0
pyaes==1.6.1
pyasn1==0.6.1
rsa==4.9
Telethon==1.39.0
Werkzeug==3.1.3
python-dotenv==1.0.1